requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0

# Azure
azure-ai-documentintelligence>=1.0.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
openai>=1.0.0

# DB
oracledb>=2.0.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data
pyarrow>=12.0.0
fastparquet>=2023.0.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.65.0
PyYAML>=6.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tiktoken>=0.5.0

# UI
gradio>=4.0.0

psycopg2-binary
